Backend/requirements.txt
Flask==3.1.2
flask_cors==6.0.1
Flask_SocketIO==5.5.1
python-socketio
python-engineio
pymysql==1.1.2
python-dotenv
eventlet
gunicorn